+++ voxygen/src/hud/mod.rs
@@ -2465,7 +2465,7 @@ impl Hud {
                 .font_size(self.fonts.cyri.scale(14))
                 .set(self.ids.debug_info, ui_widgets);
             }
-        } else {
+        } else if global_state.settings.interface.toggle_hotkey_hints {
             prof_span!("help window");
             // Help Window
             if let Some(help_key) = global_state.settings.controls.get_binding(GameInput::Help) {